refactor(gateway): move scope activity out of signals dal
